refactor(slice-45): slim experiments API and CLI presentation surfaces
Move Bayesian/stale-status helpers and Rich watch/summary output into
focused siblings so routers and Typer commands stay thin façades (#153).
