+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Slice 45 (embedding + retrieval packages)** — moved embedders/`rate_limiter` into `server/core/embedding/` and Mongo/Postgres retrievers into `server/core/retrieval/`; `sys.modules` shims keep `patch("server.core.sie_embedder.*")` / `retriever_mongo.*` working.
 - **Slice 45 (rerank package)** — moved `reranker` + `local_reranker` into `server/core/rerank/`; shim aliases keep old import/patch paths working.
 - **Slice 45 (db theme packages)** — moved ports/factory into `server/db/ports/`, Mongo adapters into `server/db/mongo/`, Postgres adapters + `schema.sql` into `server/db/postgres/`; shim aliases at old `server.db.*` paths.
+- **Slice 45 (fat-surface slim)** — extracted Bayesian/stale-status helpers to `server/api/experiments_lifecycle.py` (`experiments.py` 553→388) and CLI watch/summary presentation to `cli/display.py` (`cli/main.py` 519→246).
 - **Nightly Meterian OSS path** — `nightly.yml` Meterian job uses `oss: true` (no `METERIAN_API_TOKEN`), pins scanners to Python + Node (`cli_args`), and archives `meterian-<run>` artifacts (HTML, JUnit, SARIF, CycloneDX + CSV SBOM) for comparison with Anchore (**IMPLEMENTED**; not yet observed on a green nightly run).
 - **Shared coverage floors 95/90/95/95** (DECISIONS #142) — FE Vitest **95/90/95/95**; BE **95/90/n/a/95** via `fail_under=95` + `scripts/check_backend_coverage_floors.py` (stmts/br/lines from coverage JSON; functions n/a on coverage.py). Gap tests raised TOTAL ≈97.7%.
 - **Fair coverage metric floors** (DECISIONS #141) — FE **95/90/95/95**; BE interim policy 92/85 with `fail_under=90` — **superseded for BE by #142**.
