+++ docs/plan/DECISIONS.md
@@ -156,3 +156,4 @@
 | 150 | 2026-07-28 | 45 | Land `server/core/embedding/` + `server/core/retrieval/`; `sys.modules` shims at old flat paths | Same patch-compat pattern as pipeline (#148); factory tests retarget `sys.modules` keys to `embedding.*`; rerank deferred to #151 | Move rerank in same commit / rewrite all callers without shims |
 | 151 | 2026-07-28 | 45 | Land `server/core/rerank/` (`reranker`, `local_reranker`); `sys.modules` shims at old paths | Completes core Feature provider cluster from theme map; search imports canonical `server.core.rerank.reranker` | Fold into #150 / leave flat |
 | 152 | 2026-07-28 | 45 | Land `server/db/{ports,mongo,postgres}/`; shims at old flat `server.db.*` paths except **no** `server/db/postgres.py` (collides with package); pool API re-exported on `server.db.postgres` package; `schema.sql` co-located with `postgres.py` | Must hotspot #2; preserves Protocols + factory; `from server.db import postgres` / `close_pool` stays valid | Split db across multiple PRs / invent `pg/` package rename |
+| 153 | 2026-07-28 | 45 | Fat-surface slim: `experiments_lifecycle.py` + `cli/display.py`; keep routers/commands as façades | Execution order step 3 after db; pre→post lines experiments 553→388, cli/main 519→246 | Defer slim until FE primitives / leave god modules |
